Lần này test ver2

+++ src/red/jdbc/daos/GodGK.java
@@ -291,9 +291,9 @@ public class GodGK {
                             player.nPoint.tiemNang = Long.parseLong(String.valueOf(dataArray.get(2)));
                             player.nPoint.stamina = Short.parseShort(String.valueOf(dataArray.get(3)));
                             player.nPoint.maxStamina = Short.parseShort(String.valueOf(dataArray.get(4)));
-                            player.nPoint.hpg = Integer.parseInt(String.valueOf(dataArray.get(5)));
-                            player.nPoint.mpg = Integer.parseInt(String.valueOf(dataArray.get(6)));
-                            player.nPoint.dameg = Integer.parseInt(String.valueOf(dataArray.get(7)));
+                            player.nPoint.hpg = Long.parseLong(String.valueOf(dataArray.get(5)));
+                            player.nPoint.mpg = Long.parseLong(String.valueOf(dataArray.get(6)));
+                            player.nPoint.dameg = Long.parseLong(String.valueOf(dataArray.get(7)));
                             player.nPoint.defg = Integer.parseInt(String.valueOf(dataArray.get(8)));
                             player.nPoint.critg = Byte.parseByte(String.valueOf(dataArray.get(9)));
                             dataArray.get(10); // ** Năng động
@@ -739,13 +739,13 @@ public class GodGK {
                                 pet.nPoint.tiemNang = Long.parseLong(String.valueOf(dataArray.get(2)));
                                 pet.nPoint.stamina = Short.parseShort(String.valueOf(dataArray.get(3)));
                                 pet.nPoint.maxStamina = Short.parseShort(String.valueOf(dataArray.get(4)));
-                                pet.nPoint.hpg = Integer.parseInt(String.valueOf(dataArray.get(5)));
-                                pet.nPoint.mpg = Integer.parseInt(String.valueOf(dataArray.get(6)));
-                                pet.nPoint.dameg = Integer.parseInt(String.valueOf(dataArray.get(7)));
+                                pet.nPoint.hpg = Long.parseLong(String.valueOf(dataArray.get(5)));
+                                pet.nPoint.mpg = Long.parseLong(String.valueOf(dataArray.get(6)));
+                                pet.nPoint.dameg = Long.parseLong(String.valueOf(dataArray.get(7)));
                                 pet.nPoint.defg = Integer.parseInt(String.valueOf(dataArray.get(8)));
                                 pet.nPoint.critg = Byte.parseByte(String.valueOf(dataArray.get(9)));
-                                int hp = Integer.parseInt(String.valueOf(dataArray.get(10)));
-                                int mp = Integer.parseInt(String.valueOf(dataArray.get(11)));
+                                long hp = Long.parseLong(String.valueOf(dataArray.get(10)));
+                                long mp = Long.parseLong(String.valueOf(dataArray.get(11)));
 
                                 // data body
                                 dataArray = (JSONArray) jv.parse(String.valueOf(petData.get(2)));
@@ -874,9 +874,9 @@ public class GodGK {
                 player.nPoint.tiemNang = Long.parseLong(String.valueOf(dataArray.get(2)));
                 player.nPoint.stamina = Short.parseShort(String.valueOf(dataArray.get(3)));
                 player.nPoint.maxStamina = Short.parseShort(String.valueOf(dataArray.get(4)));
-                player.nPoint.hpg = Integer.parseInt(String.valueOf(dataArray.get(5)));
-                player.nPoint.mpg = Integer.parseInt(String.valueOf(dataArray.get(6)));
-                player.nPoint.dameg = Integer.parseInt(String.valueOf(dataArray.get(7)));
+                player.nPoint.hpg = Long.parseLong(String.valueOf(dataArray.get(5)));
+                player.nPoint.mpg = Long.parseLong(String.valueOf(dataArray.get(6)));
+                player.nPoint.dameg = Long.parseLong(String.valueOf(dataArray.get(7)));
                 player.nPoint.defg = Integer.parseInt(String.valueOf(dataArray.get(8)));
                 player.nPoint.critg = Byte.parseByte(String.valueOf(dataArray.get(9)));
                 dataArray.get(10); // ** Năng động
@@ -1010,9 +1010,9 @@ public class GodGK {
                     pet.nPoint.tiemNang = Long.parseLong(String.valueOf(dataArray.get(2)));
                     pet.nPoint.stamina = Short.parseShort(String.valueOf(dataArray.get(3)));
                     pet.nPoint.maxStamina = Short.parseShort(String.valueOf(dataArray.get(4)));
-                    pet.nPoint.hpg = Integer.parseInt(String.valueOf(dataArray.get(5)));
-                    pet.nPoint.mpg = Integer.parseInt(String.valueOf(dataArray.get(6)));
-                    pet.nPoint.dameg = Integer.parseInt(String.valueOf(dataArray.get(7)));
+                    pet.nPoint.hpg = Long.parseLong(String.valueOf(dataArray.get(5)));
+                    pet.nPoint.mpg = Long.parseLong(String.valueOf(dataArray.get(6)));
+                    pet.nPoint.dameg = Long.parseLong(String.valueOf(dataArray.get(7)));
                     pet.nPoint.defg = Integer.parseInt(String.valueOf(dataArray.get(8)));
                     pet.nPoint.critg = Byte.parseByte(String.valueOf(dataArray.get(9)));
                     int hp = Integer.parseInt(String.valueOf(dataArray.get(10)));
@@ -1099,9 +1099,9 @@ public class GodGK {
                 player.nPoint.tiemNang = Long.parseLong(String.valueOf(dataArray.get(2)));
                 player.nPoint.stamina = Short.parseShort(String.valueOf(dataArray.get(3)));
                 player.nPoint.maxStamina = Short.parseShort(String.valueOf(dataArray.get(4)));
-                player.nPoint.hpg = Integer.parseInt(String.valueOf(dataArray.get(5)));
-                player.nPoint.mpg = Integer.parseInt(String.valueOf(dataArray.get(6)));
-                player.nPoint.dameg = Integer.parseInt(String.valueOf(dataArray.get(7)));
+                player.nPoint.hpg = Long.parseLong(String.valueOf(dataArray.get(5)));
+                player.nPoint.mpg = Long.parseLong(String.valueOf(dataArray.get(6)));
+                player.nPoint.dameg = Long.parseLong(String.valueOf(dataArray.get(7)));
                 player.nPoint.defg = Integer.parseInt(String.valueOf(dataArray.get(8)));
                 player.nPoint.critg = Byte.parseByte(String.valueOf(dataArray.get(9)));
                 dataArray.get(10); // ** Năng động
